+++ test/fixtures/dev-server/empty-entry.js
@@ -7,9 +7,8 @@ module.exports = {
   stats: { orphanModules: true, preset: 'detailed' },
   entry: {},
   devServer: {
-    transportMode: {
-      server: 'sockjs',
-      client: 'sockjs',
+    client: {
+      transport: 'sockjs',
     },
   },
   plugins: [ExitOnDonePlugin],
